requirements.txt
aiohttp == 3.5.4
beautifulsoup4 == 4.8.1
discord.py == 1.2.5
feedparser == 5.2.1
html2text == 2019.9.26
peony_twitter == 1.1.5
